requirements.txt
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_884c889c-96af-444f-bd6d-daddb5e9a462ykj3l5n_/croots/recipe/certifi_1655968814730/work/certifi
charset-normalizer==2.1.0
click==8.1.3
cmake==3.22.5
contextlib2==21.6.0
Django==4.0.6
djangorestframework==3.13.1
dlib==19.23.0
execnet==1.9.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.3
imageio==2.19.3
iniconfig==1.1.1
mock==4.0.3
networkx==2.8.4
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
path==16.4.0
path.py==12.5.0
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.9
pytest==7.1.2
pytest-shutil==1.7.0
pytz==2022.1
PyWavelets==1.3.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
tifffile==2022.5.4
toml==0.10.2
tomli==2.0.1
urllib3==1.26.10
